+++ app/src/main/java/com/example/cafmealplanner/ui/Menu/MenuFragment.java
@@ -1,4 +1,4 @@
-package com.example.cafmealplanner.ui.notifications;
+package com.example.cafmealplanner.ui.Menu;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,23 +9,22 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
+import com.example.cafmealplanner.databinding.FragmentMenuBinding;
 
-import com.example.cafmealplanner.databinding.FragmentProfileBinding;
+public class MenuFragment extends Fragment {
 
-public class NotificationsFragment extends Fragment {
-
-    private FragmentProfileBinding binding;
+    private FragmentMenuBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        NotificationsViewModel notificationsViewModel =
-                new ViewModelProvider(this).get(NotificationsViewModel.class);
+        MenuViewModel menuViewModel =
+                new ViewModelProvider(this).get(MenuViewModel.class);
 
-        binding = FragmentProfileBinding.inflate(inflater, container, false);
+        binding = FragmentMenuBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textNotifications;
-        notificationsViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textHome;
+        menuViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
