+++ app/src/main/java/com/example/cafmealplanner/ui/Menu/MenuViewModel.java
@@ -1,16 +1,16 @@
-package com.example.cafmealplanner.ui.dashboard;
+package com.example.cafmealplanner.ui.Menu;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class MenuViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public MenuViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is dashboard fragment");
+        mText.setValue("This is Menu fragment");
     }
 
     public LiveData<String> getText() {
