+++ app/src/main/java/com/example/cafmealplanner/ui/Profile/ProfileViewModel.java
@@ -1,16 +1,16 @@
-package com.example.cafmealplanner.ui.notifications;
+package com.example.cafmealplanner.ui.Profile;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public ProfileViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is notifications fragment");
+        mText.setValue("This is Profile fragment");
     }
 
     public LiveData<String> getText() {
