+++ app/src/main/java/com/example/cafmealplanner/ui/Schedule/ScheduleFragment.java
@@ -1,4 +1,4 @@
-package com.example.cafmealplanner.ui.home;
+package com.example.cafmealplanner.ui.Schedule;
 
 import android.os.Bundle;
 import android.view.LayoutInflater;
@@ -9,22 +9,23 @@ import android.widget.TextView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.ViewModelProvider;
-import com.example.cafmealplanner.databinding.FragmentMenuBinding;
 
-public class HomeFragment extends Fragment {
+import com.example.cafmealplanner.databinding.FragmentScheduleBinding;
 
-    private FragmentMenuBinding binding;
+public class ScheduleFragment extends Fragment {
+
+    private FragmentScheduleBinding binding;
 
     public View onCreateView(@NonNull LayoutInflater inflater,
                              ViewGroup container, Bundle savedInstanceState) {
-        HomeViewModel homeViewModel =
-                new ViewModelProvider(this).get(HomeViewModel.class);
+        ScheduleViewModel dashboardViewModel =
+                new ViewModelProvider(this).get(ScheduleViewModel.class);
 
-        binding = FragmentMenuBinding.inflate(inflater, container, false);
+        binding = FragmentScheduleBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
 
-        final TextView textView = binding.textHome;
-        homeViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
+        final TextView textView = binding.textDashboard;
+        dashboardViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
         return root;
     }
