+++ app/src/main/java/com/example/cafmealplanner/ui/Schedule/ScheduleViewModel.java
@@ -1,16 +1,16 @@
-package com.example.cafmealplanner.ui.home;
+package com.example.cafmealplanner.ui.Schedule;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class HomeViewModel extends ViewModel {
+public class ScheduleViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public HomeViewModel() {
+    public ScheduleViewModel() {
         mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
+        mText.setValue("This is schedule fragment");
     }
 
     public LiveData<String> getText() {
